Allow comparison of String_UUID and Text
The code at
https://github.com/erezsh/sqeleton/blob/8655be43096dd6610c4ed8b5f9713f9a97670e7e/sqeleton/databases/base.py#L523-L534
uses sampling to determine if a column is String_UUID.
But when such a column is nullable, it depends on chance if the sample
contains NULL values or not. If it does, sqeleton will use the Text
type, otherwise it will use String_UUID.
Thise change ensures that such columns can be compared reliably.